+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # domino x.x.x (not yet released)
 * Fix unescape mechanism in attribute values. (#95)
 * Disable nonstandard "ignore case" version of attribute matching.
+* Add `dom/nodes` tests from w3c/web-platform-tests. (#92, @pimterry)
 
 # domino 1.0.27 (17 Oct 2016)
 * Fix bug in AFE list replacement over existing bookmark.